core: move patches and extensions to folder
